♻️ chat.service 도메인별 분할 (314 → 37+122+119줄)
- chat.service.ts (37): 컨트롤러 호환 파사드
- chat-room.service.ts (122): 채팅방 CRUD
- chat-message.service.ts (119): 메시지 송수신 + 알림 생성

모든 파일 200줄 이하 ✅
