+++ apps/chat/src/chat.module.ts
@@ -1,12 +1,14 @@
 import { Module } from '@nestjs/common';
 import { ChatController } from './chat.controller';
 import { ChatService } from './chat.service';
+import { ChatRoomService } from './chat-room.service';
+import { ChatMessageService } from './chat-message.service';
 import { ChatGateway } from './chat.gateway';
 import { PrismaModule } from '@app/prisma';
 
 @Module({
   imports: [PrismaModule],
   controllers: [ChatController],
-  providers: [ChatService, ChatGateway],
+  providers: [ChatService, ChatRoomService, ChatMessageService, ChatGateway],
 })
 export class ChatModule {}
